refactor: move core executor to core crate
