+++ router/Cargo.toml
@@ -58,3 +58,4 @@ anyhow = { workspace = true }
 prost-build = { workspace = true }
 
 [dev-dependencies]
+commonware-avs-core = { path = "../core", features = ["test-utils"] }
