+++ router/src/executor/bls/tests/mod.rs
@@ -1,5 +1,4 @@
 use crate::executor::bls::{BlsSignatureVerificationHandler, convert_non_signer_data};
-use crate::executor::core::ExecutionResult;
 use alloy_primitives::{Bytes, FixedBytes, U256};
 use anyhow::Result;
 use async_trait::async_trait;
@@ -7,6 +6,7 @@ use commonware_avs_bindings::blssigcheckoperatorstateretriever::{
     BLSSigCheckOperatorStateRetriever::getNonSignerStakesAndSignatureReturn, BN254,
     IBLSSignatureCheckerTypes,
 };
+use commonware_avs_core::traits::ExecutionResult;
 
 /// Mock implementation of BlsSignatureVerificationHandler for testing purposes.
 #[derive(Debug)]
