--- router/src/executor/core/mod.rs
@@ -1,5 +1,0 @@
-pub mod traits;
-pub mod types;
-
-pub use traits::{ExecutionResult, VerificationExecutor};
-pub use types::VerificationData;
