--- router/src/executor/core/traits.rs
@@ -1,1 +1,0 @@
-pub use commonware_avs_core::traits::{ExecutionResult, VerificationExecutor};
