--- router/src/executor/core/types.rs
@@ -1,1 +1,0 @@
-pub use commonware_avs_core::types::VerificationData;
