+++ router/src/executor/mod.rs
@@ -2,7 +2,6 @@
 
 // Public modules
 pub mod bls;
-pub mod core;
 
 // Test module
 #[cfg(test)]
@@ -11,8 +10,6 @@ pub mod tests;
 // Re-export the main types for easy access
 #[allow(unused_imports)]
 pub use bls::{BlsEigenlayerExecutor, convert_non_signer_data};
-#[allow(unused_imports)]
-pub use core::{ExecutionResult, VerificationData, VerificationExecutor};
 
 // Re-export test utilities
 #[cfg(test)]
